requirement.txt
telethon
python-dotenv
pandas
emoji
psycopg2-binary
fastapi
uvicorn[standard]
pyyaml
requests
dbt-core
ultralytics
 fastapi 
 uvicorn[standard]
 dagster
 dagster-webserver
 dagster-dg-cli